Finished v1 of multi properties
Finished the grad_config and grad_param_config.
Made some changes to grad_params so that it now returns the same result N times.
Added a Base.complex as well as an alloc_B.
We need to decide weather we want the full result as is for grad_param_config as well as changing to non-allocating code.