Changed grad_params to return SVectors
Took away the multiple alloc_B dispatching and instead added a ACEbase.valtype to handle SVectors.
Similarly I added a new genmul that works on SVectors by copying the result N times for each of the N properties.
grad_params test was commented out for now, we need a way to get a column of an SVector.